# gen/openapiv2/score_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "score_service.proto",
    "version": "0.1"
  },
  "tags": [
    {
      "name": "ScoreServiceV1"
    }
  ],
  "host": "localhost:8999",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/scores": {
      "get": {
        "summary": "List top k highest score client",
        "description": "List top k highest score client on the server",
        "operationId": "ScoreServiceV1_ListTopKScores",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoListTopKScoresV1OutterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ScoreServiceV1"
        ]
      },
      "post": {
        "summary": "Insert a client score",
        "description": "Insert a client score to the server.",
        "operationId": "ScoreServiceV1_InsertScoreV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoInsertScoreV1Response"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoScoreV1"
            }
          }
        ],
        "tags": [
          "ScoreServiceV1"
        ]
      }
    }
  },
  "definitions": {
    "protoInsertScoreV1Response": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/protoScoreV1"
        },
        "transit_id": {
          "type": "string"
        }
      }
    },
    "protoListTopKScoresV1InnerResponse": {
      "type": "object",
      "properties": {
        "scores": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protoScoreV1"
          }
        }
      }
    },
    "protoListTopKScoresV1OutterResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/protoListTopKScoresV1InnerResponse"
        },
        "transit_id": {
          "type": "string"
        }
      }
    },
    "protoScoreV1": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "@inject_tag: json:\"id\" bson:\"id\" gorm:\"column:id;primary_key\""
        },
        "created": {
          "type": "string",
          "format": "int64",
          "title": "@inject_tag: json:\"created\" bson:\"created\" gorm:\"column:created\""
        },
        "updated": {
          "type": "string",
          "format": "int64",
          "title": "@inject_tag: json:\"updated\" bson:\"updated\" gorm:\"column:updated\""
        },
        "score": {
          "type": "number",
          "format": "float",
          "title": "@inject_tag: json:\"score,omitempty\" bson:\"score\" gorm:\"column:score;index:score\""
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}